lots of updates, working on a base layout and css theme

+++ webservice/src/main/java/org/ncr/thrustcurve/v1/V1Contract.java
@@ -10,9 +10,9 @@ public interface V1Contract {
 
     Object search(MotorImpulse impulse) throws IOException;
 
-    Object search(int diameter) throws IOException;
+    Object search(float diameter) throws IOException;
 
-    List<Integer> diameters();
+    List<Float> diameters();
 
     List<String> impulses();
