+++ webservice/src/main/java/org/ncr/thrustcurve/v2/V2Endpoints.java
@@ -45,7 +45,7 @@ public class V2Endpoints implements V2Contract {
 
     @Override
     @RequestMapping(value = "/search/diameter/{diameter}", method = RequestMethod.GET, produces = "application/json")
-    public List<MotorDTO> search(@PathVariable("diameter") int diameter) throws IOException {
+    public List<MotorDTO> search(@PathVariable("diameter") float diameter) throws IOException {
         return search(new SearchCriteria().diameter(diameter), false);
     }
 
@@ -63,7 +63,7 @@ public class V2Endpoints implements V2Contract {
 
     @Override
     @RequestMapping(value = "/list/diameters", method = RequestMethod.GET, produces = "application/json")
-    public List<Integer> diameters() { return cache.getDiameters().stream().map(d -> d.getDiameter()).collect(Collectors.toList()); }
+    public List<Float> diameters() { return cache.getDiameters().stream().map(d -> d.getDiameter()).collect(Collectors.toList()); }
 
     @Override
     @RequestMapping(value = "/list/impulses", method = RequestMethod.GET, produces = "application/json")
